feat: add bg to circle xp bar
