+++ src/v2/components/ui/Avatar.tsx
@@ -1,5 +1,3 @@
-"use client"
-
 import { cn } from "@/lib/utils"
 import * as AvatarPrimitive from "@radix-ui/react-avatar"
 import { VariantProps, cva } from "class-variance-authority"
